+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Change Log
  * Fix bug handling form request rules: they can be arrays of rules
  * Fix bug where `MutableCollection` is being used, but it is not a dependency
  * Refactor `RequestArgumentHelper` to use traits so functionality can be shared
+ * Make trait properties and methods protected instead of private
 
 2021-04-14
 ----------